Fixes-fixes-fixes + different collisions
Fixed memory collisions with the spritesheet/mapsheet shared memory. Took the bottom half of the title and the shared memory is now exclusive to the map. Fixed the title draw function to accommodate the new position
Fixed collision checking, this *might* have been the issue with sprite 005 and 008 causing erroneous behavior when set as walls. Can't confirm but the error can no longer be reproduced
Expanded collision checking, now it will check for friendly and enemy buildings too. Priority order is enemy->wall-friendly. Should be easy to code different behaviors for when shot at.